arch(policy): move interface into shared
