+++ shared/policy/common/interfaces/TripInterface.ts
